Rename modules to snake_case
PEP 8 says "Modules should have short, all-lowercase names. Underscores can be used in the module name if it improves readability." I don't know about short, but at least they're all lowercase now.
